# apps/api/openapi/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FOM API",
    "description": "API for FOM backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InteractionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "stakeholder": {
            "type": "string"
          },
          "communicationDate": {
            "type": "string"
          },
          "communicationDetails": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "attachmentId": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "stakeholder",
          "communicationDate",
          "communicationDetails",
          "projectId",
          "attachmentId"
        ]
      },
      "UpdateInteractionDto": {
        "type": "object",
        "properties": {
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "stakeholder": {
            "type": "string"
          },
          "communicationDate": {
            "type": "string"
          },
          "communicationDetails": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "attachmentId": {
            "type": "number"
          }
        },
        "required": [
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "stakeholder",
          "communicationDate",
          "communicationDetails",
          "projectId",
          "attachmentId"
        ]
      },
      "DistrictDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ForestClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ProjectPublicSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "geojson": {
            "type": "object",
            "example": " { \"type\": \"Point\", \"coordinates\": [-119.396071939, 49.813816629]}"
          },
          "fspId": {
            "type": "number"
          },
          "forestClientName": {
            "type": "string"
          },
          "workflowStateName": {
            "type": "string"
          },
          "commentingOpenDate": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "geojson",
          "fspId",
          "forestClientName",
          "workflowStateName",
          "commentingOpenDate"
        ]
      },
      "WorkflowStateCode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "commentingOpenDate": {
            "type": "string"
          },
          "commentingClosedDate": {
            "type": "string"
          },
          "geojson": {
            "type": "object",
            "example": " { \"type\": \"Point\", \"coordinates\": [-119.396071939, 49.813816629]}"
          },
          "fspId": {
            "type": "number"
          },
          "districtId": {
            "type": "number"
          },
          "district": {
            "$ref": "#/components/schemas/DistrictDto"
          },
          "forestClientNumber": {
            "type": "string"
          },
          "forestClient": {
            "$ref": "#/components/schemas/ForestClientDto"
          },
          "workflowStateCode": {
            "type": "string"
          },
          "workflowState": {
            "$ref": "#/components/schemas/WorkflowStateCode"
          }
        },
        "required": [
          "id",
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "name",
          "description",
          "commentingOpenDate",
          "commentingClosedDate",
          "geojson",
          "fspId",
          "districtId",
          "district",
          "forestClientNumber",
          "forestClient",
          "workflowStateCode",
          "workflowState"
        ]
      },
...
      "ProjectSpatialDetail": {
        "type": "object",
        "properties": {
          "objectId": {
            "type": "number"
          },
          "sourceTable": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "geometry": {
            "type": "object"
          },
          "plannedDevelopmentDate": {
            "type": "string"
          },
          "plannedAreaHa": {
            "type": "number"
          },
          "plannedLengthKm": {
            "type": "number"
          },
          "submissionTypeCode": {
            "type": "string"
          },
          "submissionType": {
            "$ref": "#/components/schemas/SubmissionTypeCode"
          },
          "projectId": {
            "type": "number"
          }
        },
        "required": [
          "objectId",
          "sourceTable",
          "name",
          "geometry",
          "plannedDevelopmentDate",
          "plannedAreaHa",
          "plannedLengthKm",
          "submissionTypeCode",
          "submissionType",
          "projectId"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "commentingOpenDate": {
            "type": "string"
          },
          "commentingClosedDate": {
            "type": "string"
          },
          "geojson": {
            "type": "object",
            "example": " { \"type\": \"Point\", \"coordinates\": [-119.396071939, 49.813816629]}"
          },
          "fspId": {
            "type": "number"
          },
          "districtId": {
            "type": "number"
          },
          "district": {
            "$ref": "#/components/schemas/DistrictDto"
          },
          "forestClientNumber": {
            "type": "string"
          },
          "forestClient": {
            "$ref": "#/components/schemas/ForestClientDto"
          },
          "workflowStateCode": {
            "type": "string"
          },
          "workflowState": {
            "$ref": "#/components/schemas/WorkflowStateCode"
          }
        },
        "required": [
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "name",
          "description",
          "commentingOpenDate",
          "commentingClosedDate",
          "geojson",
          "fspId",
          "districtId",
          "district",
          "forestClientNumber",
          "forestClient",
          "workflowStateCode",
          "workflowState"
        ]
      },
      "DeleteResult": {
        "type": "object",
        "properties": {}
      },
      "Project": {
        "type": "object",
        "properties": {}
      },
...
      "PublicCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "feedback": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "responseDetails": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "responseCode": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/ResponseCode"
          },
          "commentScopeCode": {
            "type": "string"
          },
          "commentScope": {
            "$ref": "#/components/schemas/CommentScopeCode"
          },
          "scopeCutBlockId": {
            "type": "number"
          },
          "scopeRoadSectionId": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "feedback",
          "name",
          "location",
          "email",
          "phoneNumber",
          "responseDetails",
          "projectId",
          "project",
          "responseCode",
          "response",
          "commentScopeCode",
          "commentScope",
          "scopeCutBlockId",
          "scopeRoadSectionId"
        ]
      },
      "UpdatePublicCommentDto": {
        "type": "object",
        "properties": {
          "revisionCount": {
            "type": "number"
          },
          "createTimestamp": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "updateTimestamp": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          },
          "feedback": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "responseDetails": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "responseCode": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/ResponseCode"
          },
          "commentScopeCode": {
            "type": "string"
          },
          "commentScope": {
            "$ref": "#/components/schemas/CommentScopeCode"
          },
          "scopeCutBlockId": {
            "type": "number"
          },
          "scopeRoadSectionId": {
            "type": "number"
          }
        },
        "required": [
          "revisionCount",
          "createTimestamp",
          "createUser",
          "updateTimestamp",
          "updateUser",
          "feedback",
          "name",
          "location",
          "email",
          "phoneNumber",
          "responseDetails",
          "projectId",
          "project",
          "responseCode",
          "response",
          "commentScopeCode",
          "commentScope",
          "scopeCutBlockId",
          "scopeRoadSectionId"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/interaction": {
      "post": {
        "operationId": "InteractionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "interaction"
        ]
      }
    },
    "/api/interaction/{id}": {
      "get": {
        "operationId": "InteractionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interaction"
        ]
      },
      "put": {
        "operationId": "InteractionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInteractionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interaction"
        ]
      },
      "delete": {
        "operationId": "InteractionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interaction"
        ]
      }
    },
...
    "/api/forest-client/{id}": {
      "get": {
        "operationId": "ForestClientController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForestClientDto"
                }
              }
            }
          }
        },
        "tags": [
          "forest-client"
        ]
      }
    },
    "/api/project/publicSummary": {
      "get": {
        "operationId": "ProjectController_findPublicSummary",
        "parameters": [
          {
            "name": "includeCommentOpen",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includePostCommentOpen",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forestClientName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openedOnOrAfter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectPublicSummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/api/project": {
      "get": {
        "operationId": "ProjectController_find",
        "parameters": [
          {
            "name": "fspId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "districtId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "workflowStateCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forestClientName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/api/project/spatialDetails/{id}": {
      "get": {
        "operationId": "ProjectController_getSpatialDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSpatialDetail"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/api/project/{id}": {
      "get": {
        "operationId": "ProjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      },
      "put": {
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      },
      "delete": {
        "operationId": "ProjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          }
        },
        "tags": [
          "project"
        ]
      }
    },
...
    "/api/public-comment": {
      "post": {
        "operationId": "PublicCommentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "public-comment"
        ]
      },
      "get": {
        "operationId": "PublicCommentController_find",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicCommentDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "public-comment"
        ]
      }
    },
    "/api/public-comment/byProjectId/{id}": {
      "get": {
        "operationId": "PublicCommentController_findByProjectId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-comment"
        ]
      }
    },
    "/api/public-comment/{id}": {
      "get": {
        "operationId": "PublicCommentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-comment"
        ]
      },
      "put": {
        "operationId": "PublicCommentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublicCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-comment"
        ]
      },
      "delete": {
        "operationId": "PublicCommentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-comment"
        ]
      }
    },
...
  }
}